patient_module/requirements.txt
google-cloud-bigquery
requests
openai
python-dotenv


